Include monitoring exporter sidecar and Prometheus annotations for admin server
